refactor: extract binary utilities to resolve max-lines lint error
